Fixed stencil masking
This is odd, but stencil masking was simply disabled, so
clearing the stencil buffer would only work if the mask
was 0xFF, or it would simply skip the clearing entirely.
I removed this condition to fix the issue. Also removed
some dead code and added an early exit if the mask is 0.

Change-Id: I359b10ed3382b75cb9d078470f237e68f1a6e7b9
Reviewed-on: https://swiftshader-review.googlesource.com/4303
Tested-by: Alexis Hétu <sugoi@google.com>
Reviewed-by: Nicolas Capens <capn@google.com>