Fixed BindRenderbuffer
BindRenderbuffer doesn't require anything except a non-zero
renderbuffer to create a new Renderbuffer object (it is not
required to use GenRenderbuffers before using BindRenderbuffer).
This had been fixed for GLES 3.0, but the same also applies
to GLES 2.0, which has been fixed here. Note that the Context
already handles the case where renderbuffer is zero properly.

Bug 25983901

Change-Id: I75f706fc56df28871bf60b5645f1c0db039fccb8
Reviewed-on: https://swiftshader-review.googlesource.com/4334
Tested-by: Alexis Hétu <sugoi@google.com>
Reviewed-by: Nicolas Capens <capn@google.com>