Support HAL_PIXEL_FORMAT_IMPLEMENTATION_DEFINED as RGBX8888
The 3.0 camera HAL makes extensive use of
HAL_PIXEL_FORMAT_IMPLEMENTATION_DEFINED, which Android defines as the
format that is best for the hardware. The fake camera implementation
decided that that was RGBX8888, so it seem like that is what we should
use.

Test: untested
Change-Id: Ia02ebbefd015f9ca099b4393daf5d464f38fe8fc
Reviewed-on: https://swiftshader-review.googlesource.com/8588
Reviewed-by: Nicolas Capens <capn@google.com>
Tested-by: Nicolas Capens <capn@google.com>
(cherry picked from commit efe1ce9d263a0786fc4db64f316f57288e3fef3b)
Reviewed-on: https://swiftshader-review.googlesource.com/8610